Add tests

+++ algorithms/sorting_algorithms/src/heap_sort_lib/Heap.hpp
@@ -15,7 +15,7 @@ class Heap {
 public:
     Heap() = delete;
 
-    explicit Heap(const size_t capa)
+    explicit Heap(const int capa)
         : arr(new std::pair<K, V>[capa]), size(0), capacity(capa)
     {
     }
@@ -35,19 +35,19 @@ public:
     std::pair<K, V> extract_max();
 
 private:
-    size_t find_index(const K& key) const;
-    void heapify_up(const size_t index);
-    void heapify_down(size_t index);
+    int find_index(const K& key) const;
+    void heapify_up(const int index);
+    void heapify_down(int index);
 
     std::pair<K, V>* const arr;
-    size_t size;
-    size_t capacity;
+    int size;
+    int capacity;
 };
 
 template <typename K, typename V>
-size_t Heap<K, V>::find_index(const K& key) const
+int Heap<K, V>::find_index(const K& key) const
 {
-    size_t i = 0;
+    int i = 0;
     for (; i < size; ++i) {
         if (key == arr[i].first) {
             break;
@@ -57,21 +57,20 @@ size_t Heap<K, V>::find_index(const K& key) const
 }
 
 template <typename K, typename V>
-void Heap<K, V>::heapify_up(const size_t index)
+void Heap<K, V>::heapify_up(const int index)
 {
-    for (size_t i = index; i > 0 && arr[(i - 1) / 2] < arr[i];
-         i = (i - 1) / 2) {
+    for (int i = index; i > 0 && arr[(i - 1) / 2] < arr[i]; i = (i - 1) / 2) {
         std::swap(arr[i], arr[(i - 1) / 2]);
     }
 }
 
 template <typename K, typename V>
-void Heap<K, V>::heapify_down(size_t index)
+void Heap<K, V>::heapify_down(int index)
 {
     while (1) {
-        size_t left = 2 * index + 1;
-        size_t right = 2 * index + 2;
-        size_t largest = index;
+        int left = 2 * index + 1;
+        int right = 2 * index + 2;
+        int largest = index;
         if (left <= size && arr[left].first > arr[largest].first)
             largest = left;
         if (right <= size && arr[right].first > arr[largest].first)
@@ -118,7 +117,7 @@ void Heap<K, V>::increase_key(const K& key, const K& newkey)
         return;
     }
 
-    size_t index = find_index(key);
+    int index = find_index(key);
 
     arr[index].first = newkey;
 
@@ -146,7 +145,7 @@ std::pair<K, V> Heap<K, V>::extract_max()
 template <typename K, typename V>
 std::ostream& operator<<(std::ostream& os, const Heap<K, V>& heap)
 {
-    for (size_t c = 0; c != heap.size; ++c) {
+    for (int c = 0; c != heap.size; ++c) {
         os << heap.arr[c].first << ' ' << heap.arr[c].second << '\n';
     }
     return os;
@@ -158,7 +157,7 @@ bool operator==(const Heap<KT, VT>& h1, const Heap<KT, VT>& h2)
     if (h1.size != h2.size) {
         return 0;
     }
-    for (size_t c1 = 0, c2 = 0; c1 < h1.size && c2 < h2.size; ++c1, ++c2) {
+    for (int c1 = 0, c2 = 0; c1 < h1.size && c2 < h2.size; ++c1, ++c2) {
         if (h1.arr[c1] != h2.arr[c2]) {
             return 0;
         }
